+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 2.8.0-alpha.pre
 -----------------
+ * Removed authclient. In a separate package skeeks/cms-authclient
  * Change component connect admin menu
  * Removed http auth
  * Removed DescriptionFullColumn, DescriptionShortColumn